Remove duplicate test cases
