+++ src/test/java/com/fasterxml/jackson/databind/deser/filter/NullConversionsPojoTest.java
@@ -195,44 +195,5 @@ public class NullConversionsPojoTest extends BaseMapTest
         assertEquals(actual.name, "Computer");
         assertTrue(actual.prices.isEmpty());
     }
-
-    static class Issue3645BeanB {
-        private String text;
-        private Collection<Integer> products;
-
-        public Issue3645BeanB(
-            @JsonProperty("text") String text,
-            @JsonProperty("products") Collection<Integer>products
-        ) {
-            this.text = text;
-            this.products = products;
-        }
-    }
-
-    public void testDeserializeWithNullAsEmptyWithBuilderDefaults() throws JsonProcessingException {
-        String json = "{\"text\": \"Computer\", \"products\" : null}";
-
-        ObjectMapper mapper = JsonMapper.builder()
-            .defaultSetterInfo(JsonSetter.Value.construct(Nulls.AS_EMPTY, Nulls.AS_EMPTY))
-            .build();
-
-        Issue3645BeanB actual = mapper.readValue(json, Issue3645BeanB.class);
-
-        assertEquals(actual.text, "Computer");
-        assertTrue(actual.products.isEmpty());
-
-    }
-
-    public void testDeserializeWithNullAsEmptyWithBuilderDefaults2() throws JsonProcessingException {
-        String json = "{\"text\": \"Computer\"}";
-
-        ObjectMapper mapper = JsonMapper.builder()
-            .defaultSetterInfo(JsonSetter.Value.construct(Nulls.AS_EMPTY, Nulls.AS_EMPTY))
-            .build();
-
-        Issue3645BeanB actual = mapper.readValue(json, Issue3645BeanB.class);
-
-        assertEquals(actual.text, "Computer");
-        assertTrue(actual.products.isEmpty());
-    }
+    
 }
